refactor user/team loading into hook
